Broadcast multiple messages to one wallet

+++ validator-engine/validator-engine.hpp
@@ -224,7 +224,7 @@ class ValidatorEngine : public td::actor::Actor {
   std::map<CI_key, td::uint32> control_permissions_;
 
   double state_ttl_ = 0;
-  double max_mempool_num_ = 0;
+  size_t max_mempool_num_ = 0;
   double block_ttl_ = 0;
   double sync_ttl_ = 0;
   double archive_ttl_ = 0;
@@ -288,7 +288,7 @@ class ValidatorEngine : public td::actor::Actor {
   void set_state_ttl(double t) {
     state_ttl_ = t;
   }
-  void set_max_mempool_num(double t) {
+  void set_max_mempool_num(size_t t) {
     max_mempool_num_ = t;
   }
   void set_block_ttl(double t) {
