+++ validator/validator-options.hpp
@@ -49,7 +49,7 @@ struct ValidatorManagerOptionsImpl : public ValidatorManagerOptions {
   double state_ttl() const override {
     return state_ttl_;
   }
-  double max_mempool_num() const override {
+  size_t max_mempool_num() const override {
     return max_mempool_num_;
   }
   double archive_ttl() const override {
@@ -198,7 +198,7 @@ struct ValidatorManagerOptionsImpl : public ValidatorManagerOptions {
   void set_state_ttl(double value) override {
     state_ttl_ = value;
   }
-  void set_max_mempool_num(double value) override {
+  void set_max_mempool_num(size_t value) override {
     max_mempool_num_ = value;
   }
   void set_archive_ttl(double value) override {
@@ -307,7 +307,7 @@ struct ValidatorManagerOptionsImpl : public ValidatorManagerOptions {
   }
 
   ValidatorManagerOptionsImpl(BlockIdExt zero_block_id, BlockIdExt init_block_id, bool allow_blockchain_init,
-                              double sync_blocks_before, double block_ttl, double state_ttl, double max_mempool_num,
+                              double sync_blocks_before, double block_ttl, double state_ttl, size_t max_mempool_num,
                               double archive_ttl, double key_proof_ttl, bool initial_sync_disabled)
       : zero_block_id_(zero_block_id)
       , init_block_id_(init_block_id)
@@ -329,7 +329,7 @@ struct ValidatorManagerOptionsImpl : public ValidatorManagerOptions {
   double sync_blocks_before_;
   double block_ttl_;
   double state_ttl_;
-  double max_mempool_num_;
+  size_t max_mempool_num_;
   double archive_ttl_;
   double key_proof_ttl_;
   bool initial_sync_disabled_;
